+++ app/src/main/jni/genMercator.c
@@ -1,6 +1,6 @@
 /*
 
-    Copyright 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022, 2023  joshua.tee@gmail.com
+    Copyright 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022, 2023, 2024  joshua.tee@gmail.com
 
     This file is part of wX.
 
@@ -38,7 +38,7 @@ JNIEXPORT void JNICALL Java_joshuatee_wx_Jni_genMercator(
     double W_180_DIV_PI = 180.0 / M_PI;
     double W_PI_DIV_360 = M_PI / 360.0;
     double W_PI_DIV_4 = M_PI / 4.0;
-    for (int i_count = 0; i_count < count; i_count = i_count + 2) {
+    for (int i_count = 0; i_count < count; i_count += 2) {
         oBuff[i_count + 1] = -1.0f * (-1.0 * (((W_180_DIV_PI * log(tan(W_PI_DIV_4 + iBuff[i_count] * W_PI_DIV_360)))
 									   - (W_180_DIV_PI * log(tan(W_PI_DIV_4 + centerX * W_PI_DIV_360)))) * oneDegreeScaleFactor) + yImageCenterPixels);
         oBuff[i_count] = -1.0 * ((iBuff[i_count + 1] - centerY) * oneDegreeScaleFactor) + xImageCenterPixels;